feat: frontend improvements
- use real session data in components
- add signout button
- separate settings and profile pages
